refactor(frontend): remove HDN
