+++ packages/code-du-travail-data/dataset/fiches_service_public/package.json
@@ -8,6 +8,8 @@
   },
   "dependencies": {
     "@socialgouv/fiches-vdd": "1.0.0",
+    "@socialgouv/kali-data": "^1.2.18",
+    "@socialgouv/legi-data": "^1.1.1",
     "adm-zip": "0.4.13",
     "axios": "^0.19.0",
     "lodash.uniqby": "^4.7.0",
