+++ packages/code-du-travail-data/package.json
@@ -26,7 +26,7 @@
   "dependencies": {
     "@cdt/sources": "^3.3.1",
     "@elastic/elasticsearch": "^7.4.0",
-    "@socialgouv/kali-data": "1.2.18",
+    "@socialgouv/kali-data": "^1.2.18",
     "@socialgouv/legi-data": "^1.1.1",
     "esm": "^3.2.25",
     "unist-util-find": "^1.0.1",
