+++ packages/code-du-travail-frontend/src/common/ReferencesJuridiques/__tests__/ReferencesJuridiques.test.js
