+++ packages/code-du-travail-frontend/src/common/ReferencesJuridiques/typeReference.js
